coffee-service/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
sqlalchemy[asyncio]==2.0.31
asyncpg==0.29.0
pydantic-settings==2.3.4
pydantic[email]==2.8.2
python-dotenv==1.0.1
boto3==1.34.144
httpx==0.27.0
openai==1.35.0
python-jose[cryptography]==3.3.0

# dev / test
pytest==8.3.2
pytest-asyncio==0.23.8
pytest-cov==5.0.0
